fix: commit code suggestions
